mqtt/requirements.txt
opencv-python
paho-mqtt
numpy
pillow